+++ ViewModels/EppAttrFieldViewModel.cs
@@ -12,7 +12,7 @@ namespace AFBA.EPP.ViewModels
         public string DisplyAttrNm { get; set; }
         public long AttrId { get; set; }
         public bool? RqdFlg { get; set; }
-        public int? ClmnOrdr { get; set; }
+        public long? ClmnOrdr { get; set; }
         public long GrpprdctId { get; set; }
     }
 }
